Rename block_horizontal_scratch to block_svg
